Introduce a interface representing the manager
At this point we are still using the hard coded manager which allows us
to keep all calling code unchanged. The ResultAggregator constructor is
overloaded to allow the test code to take a separate path.

We need to rely on test coverage at a higher level to prove our refactoring
still holds together before we switch calling code to inject a
ResultFetcher instance